+++ mav_msgs/include/mav_msgs/conversions.h
@@ -3,6 +3,7 @@
  * Copyright 2015 Michael Burri, ASL, ETH Zurich, Switzerland
  * Copyright 2015 Markus Achtelik, ASL, ETH Zurich, Switzerland
  * Copyright 2015 Helen Oleynikova, ASL, ETH Zurich, Switzerland
+ * Copyright 2015 Mina Kamel, ASL, ETH Zurich, Switzerland
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,7 +23,6 @@
 #ifndef MAV_MSGS_CONVERSIONS_H
 #define MAV_MSGS_CONVERSIONS_H
 
-#include <Eigen/StdVector>
 #include <geometry_msgs/Point.h>
 #include <geometry_msgs/PoseStamped.h>
 #include <geometry_msgs/Quaternion.h>
@@ -34,12 +34,18 @@
 
 #include "mav_msgs/Actuators.h"
 #include "mav_msgs/AttitudeThrust.h"
+<<<<<<< HEAD
 #include "mav_msgs/common.h"
 #include "mav_msgs/default_values.h"
 #include "mav_msgs/eigen_mav_msgs.h"
+=======
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
 #include "mav_msgs/RateThrust.h"
 #include "mav_msgs/RollPitchYawrateThrust.h"
 #include "mav_msgs/TorqueThrust.h"
+#include "mav_msgs/common.h"
+#include "mav_msgs/default_values.h"
+#include "mav_msgs/eigen_mav_msgs.h"
 
 namespace mav_msgs {
 
@@ -112,6 +118,7 @@ inline void eigenOdometryFromMsg(const nav_msgs::Odometry& msg,
   odometry->velocity_B = mav_msgs::vector3FromMsg(msg.twist.twist.linear);
   odometry->angular_velocity_B =
       mav_msgs::vector3FromMsg(msg.twist.twist.angular);
+<<<<<<< HEAD
 }
 
 inline void eigenTrajectoryPointFromTransformMsg(
@@ -129,6 +136,60 @@ inline void eigenTrajectoryPointFromTransformMsg(
   trajectory_point->acceleration_W.setZero();
   trajectory_point->jerk_W.setZero();
   trajectory_point->snap_W.setZero();
+=======
+  odometry->pose_covariance_ =
+      Eigen::Map<const Eigen::Matrix<double, 6, 6>>(msg.pose.covariance.data());
+  odometry->twist_covariance_ = Eigen::Map<const Eigen::Matrix<double, 6, 6>>(
+      msg.twist.covariance.data());
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
+}
+
+inline void eigenTrajectoryPointFromTransformMsg(
+    const geometry_msgs::TransformStamped& msg,
+    EigenTrajectoryPoint* trajectory_point) {
+  assert(trajectory_point != NULL);
+
+  ros::Time timestamp = msg.header.stamp;
+
+  trajectory_point->timestamp_ns = timestamp.toNSec();
+<<<<<<< HEAD
+  trajectory_point->position_W = vector3FromPointMsg(msg.pose.position);
+  trajectory_point->orientation_W_B = quaternionFromMsg(msg.pose.orientation);
+=======
+  trajectory_point->position_W = vector3FromMsg(msg.transform.translation);
+  trajectory_point->orientation_W_B = quaternionFromMsg(msg.transform.rotation);
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
+  trajectory_point->velocity_W.setZero();
+  trajectory_point->angular_velocity_W.setZero();
+  trajectory_point->acceleration_W.setZero();
+  trajectory_point->angular_acceleration_W.setZero();
+  trajectory_point->jerk_W.setZero();
+  trajectory_point->snap_W.setZero();
+}
+
+<<<<<<< HEAD
+/**
+ * \brief Computes the MAV state (position, velocity, attitude, angular
+ * velocity) from the flat state.
+ *        Additionally, computes the acceleration expressed in body coordinates,
+ * i.e. the acceleration
+ *        measured by the IMU. No air-drag is assumed here.
+=======
+// 2 versions of this: one for PoseStamped (which fills in the timestamps
+// correctly and should be used if at all possible), and one for a raw pose
+// message.
+inline void eigenTrajectoryPointFromPoseMsg(
+    const geometry_msgs::Pose& msg, EigenTrajectoryPoint* trajectory_point) {
+  assert(trajectory_point != NULL);
+
+  trajectory_point->position_W = vector3FromPointMsg(msg.position);
+  trajectory_point->orientation_W_B = quaternionFromMsg(msg.orientation);
+  trajectory_point->velocity_W.setZero();
+  trajectory_point->angular_velocity_W.setZero();
+  trajectory_point->acceleration_W.setZero();
+  trajectory_point->angular_acceleration_W.setZero();
+  trajectory_point->jerk_W.setZero();
+  trajectory_point->snap_W.setZero();
 }
 
 inline void eigenTrajectoryPointFromPoseMsg(
@@ -139,25 +200,20 @@ inline void eigenTrajectoryPointFromPoseMsg(
   ros::Time timestamp = msg.header.stamp;
 
   trajectory_point->timestamp_ns = timestamp.toNSec();
-  trajectory_point->position_W = vector3FromPointMsg(msg.pose.position);
-  trajectory_point->orientation_W_B = quaternionFromMsg(msg.pose.orientation);
-  trajectory_point->velocity_W.setZero();
-  trajectory_point->angular_velocity_W.setZero();
-  trajectory_point->acceleration_W.setZero();
-  trajectory_point->jerk_W.setZero();
-  trajectory_point->snap_W.setZero();
+  eigenTrajectoryPointFromPoseMsg(msg.pose, trajectory_point);
 }
 
 /**
  * \brief Computes the MAV state (position, velocity, attitude, angular
- * velocity) from the flat state.
- *        Additionally, computes the acceleration expressed in body coordinates,
- * i.e. the acceleration
- *        measured by the IMU. No air-drag is assumed here.
+ * velocity, angular acceleration) from the flat state.
+ * Additionally, computes the acceleration expressed in body coordinates,
+ * i.e. the acceleration measured by the IMU. No air-drag is assumed here.
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
  *
  * \param[in] acceleration Acceleration of the MAV, expressed in world
  * coordinates.
  * \param[in] jerk Jerk of the MAV, expressed in world coordinates.
+<<<<<<< HEAD
  * \param[in] yaw Yaw angle of the MAV, expressed in world coordinates.
  * \param[in] yaw_rate Yaw rate, expressed in world coordinates.
  * \param[in] magnitude_of_gravity Magnitude of the gravity vector.
@@ -173,17 +229,50 @@ void EigenMavStateFromEigenTrajectoryPoint(
     double yaw, double yaw_rate, double magnitude_of_gravity,
     Eigen::Quaterniond* orientation, Eigen::Vector3d* acceleration_body,
     Eigen::Vector3d* angular_velocity_body);
+=======
+ * \param[in] snap Snap of the MAV, expressed in world coordinates.
+ * \param[in] yaw Yaw angle of the MAV, expressed in world coordinates.
+ * \param[in] yaw_rate Yaw rate, expressed in world coordinates.
+ * \param[in] yaw_acceleration Yaw acceleration, expressed in world coordinates.
+ * \param[in] magnitude_of_gravity Magnitude of the gravity vector.
+ * \param[out] orientation Quaternion representing the attitude of the MAV.
+ * \param[out] acceleration_body Acceleration expressed in body coordinates,
+ * i.e. the acceleration usually by the IMU.
+ * \param[out] angular_velocity_body Angular velocity of the MAV, expressed in
+ * body coordinates.
+ * \param[out] angular_acceleration_body Angular acceleration of the MAV,
+ * expressed in body coordinates.
+ */
+void EigenMavStateFromEigenTrajectoryPoint(
+    const Eigen::Vector3d& acceleration, const Eigen::Vector3d& jerk,
+    const Eigen::Vector3d& snap, double yaw, double yaw_rate,
+    double yaw_acceleration, double magnitude_of_gravity,
+    Eigen::Quaterniond* orientation, Eigen::Vector3d* acceleration_body,
+    Eigen::Vector3d* angular_velocity_body,
+    Eigen::Vector3d* angular_acceleration_body);
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
 
 /// Convenience function with default value for the magnitude of the gravity
 /// vector.
 inline void EigenMavStateFromEigenTrajectoryPoint(
     const Eigen::Vector3d& acceleration, const Eigen::Vector3d& jerk,
+<<<<<<< HEAD
     double yaw, double yaw_rate, Eigen::Quaterniond* orientation,
     Eigen::Vector3d* acceleration_body,
     Eigen::Vector3d* angular_velocity_body) {
   EigenMavStateFromEigenTrajectoryPoint(
       acceleration, jerk, yaw, yaw_rate, kGravity, orientation,
       acceleration_body, angular_velocity_body);
+=======
+    const Eigen::Vector3d& snap, double yaw, double yaw_rate,
+    double yaw_acceleration, Eigen::Quaterniond* orientation,
+    Eigen::Vector3d* acceleration_body, Eigen::Vector3d* angular_velocity_body,
+    Eigen::Vector3d* angular_acceleration_body) {
+  EigenMavStateFromEigenTrajectoryPoint(
+      acceleration, jerk, snap, yaw, yaw_rate, yaw_acceleration, kGravity,
+      orientation, acceleration_body, angular_velocity_body,
+      angular_acceleration_body);
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
 }
 
 /// Convenience function with EigenTrajectoryPoint as input and EigenMavState as
@@ -193,10 +282,18 @@ inline void EigenMavStateFromEigenTrajectoryPoint(
     EigenMavState* mav_state) {
   assert(mav_state != NULL);
   EigenMavStateFromEigenTrajectoryPoint(
+<<<<<<< HEAD
       flat_state.acceleration_W, flat_state.jerk_W, flat_state.getYaw(),
       flat_state.getYawRate(), magnitude_of_gravity,
       &(mav_state->orientation_W_B), &(mav_state->acceleration_B),
       &(mav_state->angular_velocity_B));
+=======
+      flat_state.acceleration_W, flat_state.jerk_W, flat_state.snap_W,
+      flat_state.getYaw(), flat_state.getYawRate(), flat_state.getYawAcc(),
+      magnitude_of_gravity, &(mav_state->orientation_W_B),
+      &(mav_state->acceleration_B), &(mav_state->angular_velocity_B),
+      &(mav_state->angular_acceleration_B));
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
   mav_state->position_W = flat_state.position_W;
   mav_state->velocity_W = flat_state.velocity_W;
 }
@@ -213,9 +310,22 @@ inline void EigenMavStateFromEigenTrajectoryPoint(
 
 inline void EigenMavStateFromEigenTrajectoryPoint(
     const Eigen::Vector3d& acceleration, const Eigen::Vector3d& jerk,
+<<<<<<< HEAD
     double yaw, double yaw_rate, double magnitude_of_gravity,
     Eigen::Quaterniond* orientation, Eigen::Vector3d* acceleration_body,
     Eigen::Vector3d* angular_velocity_body) {
+=======
+    const Eigen::Vector3d& snap, double yaw, double yaw_rate,
+    double yaw_acceleration, double magnitude_of_gravity,
+    Eigen::Quaterniond* orientation, Eigen::Vector3d* acceleration_body,
+    Eigen::Vector3d* angular_velocity_body,
+    Eigen::Vector3d* angular_acceleration_body) {
+  // Mapping from flat state to full state following to Mellinger [1]:
+  // See [1]: Mellinger, Daniel Warren. "Trajectory generation and control for
+  //          quadrotors." (2012), Phd-thesis, p. 15ff.
+  // http://repository.upenn.edu/cgi/viewcontent.cgi?article=1705&context=edissertations
+  //
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
   //  zb = acc+[0 0 magnitude_of_gravity]';
   //  thrust =  norm(zb);
   //  zb = zb / thrust;
@@ -229,14 +339,25 @@ inline void EigenMavStateFromEigenTrajectoryPoint(
   //
   //  q(:,i) = rot2quat([xb yb zb]);
   //
+<<<<<<< HEAD
   //  h_w = 1/thrust*acc_dot);
+=======
+  //  h_w = 1/thrust*(acc_dot - zb' * acc_dot * zb;
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
   //
   //  w(1,i) = -h_w'*yb;
   //  w(2,i) = h_w'*xb;
   //  w(3,i) = yaw_dot*[0 0 1]*zb;
 
+<<<<<<< HEAD
   assert(acceleration_body != nullptr);
   assert(angular_velocity_body != nullptr);
+=======
+  assert(orientation != nullptr);
+  assert(acceleration_body != nullptr);
+  assert(angular_velocity_body != nullptr);
+  assert(angular_acceleration_body != nullptr);
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
 
   Eigen::Vector3d xb;
   Eigen::Vector3d yb;
@@ -247,20 +368,43 @@ inline void EigenMavStateFromEigenTrajectoryPoint(
   const double inv_thrust = 1.0 / thrust;
   zb = zb * inv_thrust;
 
+<<<<<<< HEAD
   yb = zb.cross(Eigen::Vector3d(cos(yaw), sin(yaw), 0));
+=======
+  yb = zb.cross(Eigen::Vector3d(cos(yaw), sin(yaw), 0.0));
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
   yb.normalize();
 
   xb = yb.cross(zb);
 
   const Eigen::Matrix3d R((Eigen::Matrix3d() << xb, yb, zb).finished());
 
+<<<<<<< HEAD
   const Eigen::Vector3d h_w = inv_thrust * jerk;
+=======
+  const Eigen::Vector3d h_w =
+      inv_thrust * (jerk - double(zb.transpose() * jerk) * zb);
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
 
   *orientation = Eigen::Quaterniond(R);
   *acceleration_body = R.transpose() * zb * thrust;
   (*angular_velocity_body)[0] = -h_w.transpose() * yb;
   (*angular_velocity_body)[1] = h_w.transpose() * xb;
   (*angular_velocity_body)[2] = yaw_rate * zb[2];
+<<<<<<< HEAD
+=======
+
+  // Calculate angular accelerations.
+  const Eigen::Vector3d wcrossz = (*angular_velocity_body).cross(zb);
+  const Eigen::Vector3d wcrosswcrossz = (*angular_velocity_body).cross(wcrossz);
+  const Eigen::Vector3d h_a =
+      inv_thrust * (snap - double(zb.transpose() * snap) * zb) - 2 * wcrossz -
+      wcrosswcrossz + double(zb.transpose() * wcrosswcrossz) * zb;
+
+  (*angular_acceleration_body)[0] = -h_a.transpose() * yb;
+  (*angular_acceleration_body)[1] = h_a.transpose() * xb;
+  (*angular_acceleration_body)[2] = yaw_acceleration * zb[2];
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
 }
 
 inline void eigenTrajectoryPointFromMsg(
@@ -295,8 +439,14 @@ inline void eigenTrajectoryPointFromMsg(
   if (msg.accelerations.size() > 0) {
     trajectory_point->acceleration_W =
         vector3FromMsg(msg.accelerations[0].linear);
+<<<<<<< HEAD
+=======
+    trajectory_point->angular_acceleration_W =
+        vector3FromMsg(msg.accelerations[0].angular);
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
   } else {
     trajectory_point->acceleration_W.setZero();
+    trajectory_point->angular_acceleration_W.setZero();
   }
   trajectory_point->jerk_W.setZero();
   trajectory_point->snap_W.setZero();
@@ -383,7 +533,9 @@ inline void msgOdometryFromEigen(const EigenOdometry& odometry,
                                  nav_msgs::Odometry* msg) {
   assert(msg != NULL);
 
-  msg->header.stamp.fromNSec(odometry.timestamp_ns);
+  if (odometry.timestamp_ns >= 0) {
+    msg->header.stamp.fromNSec(odometry.timestamp_ns);
+  }
   pointEigenToMsg(odometry.position_W, &msg->pose.pose.position);
   quaternionEigenToMsg(odometry.orientation_W_B, &msg->pose.pose.orientation);
 
@@ -395,7 +547,13 @@ inline void msgOdometryFromEigen(const EigenOdometry& odometry,
 inline void msgPoseStampedFromEigenTrajectoryPoint(
     const EigenTrajectoryPoint& trajectory_point,
     geometry_msgs::PoseStamped* msg) {
+<<<<<<< HEAD
   msg->header.stamp.fromNSec(trajectory_point.timestamp_ns);
+=======
+  if (trajectory_point.timestamp_ns >= 0) {
+    msg->header.stamp.fromNSec(trajectory_point.timestamp_ns);
+  }
+>>>>>>> cf206701bb06a73c486af617f870c383da7d607f
   pointEigenToMsg(trajectory_point.position_W, &msg->pose.position);
   quaternionEigenToMsg(trajectory_point.orientation_W_B,
                        &msg->pose.orientation);